feature: moved ApplicationContextHolder to core module
